+++ orix/quaternion/quaternion.py
@@ -297,9 +297,9 @@ class Quaternion(Object3d):
         q = cls.from_neo_euler(axangle).unit
         return q
 
-    # TODO: Remove decorator, **kwargs, and use of "convention" in 1.0
+    # TODO: Remove decorator, **kwargs, and use of "convention" in 0.13
     @classmethod
-    @deprecated_argument("convention", "0.9", "1.0", "direction")
+    @deprecated_argument("convention", "0.9", "0.13", "direction")
     def from_euler(
         cls,
         euler: Union[np.ndarray, tuple, list],
@@ -772,8 +772,8 @@ class Quaternion(Object3d):
                 "with `other` of type `Quaternion` or `Vector3d`"
             )
 
-    # TODO: Remove decorator and **kwargs in 1.0
-    @deprecated_argument("convention", since="0.9", removal="1.0")
+    # TODO: Remove decorator and **kwargs in 0.13
+    @deprecated_argument("convention", since="0.9", removal="0.13")
     def to_euler(self, degrees: bool = False, **kwargs) -> np.ndarray:
         r"""Return the normalized quaternions as Euler angles in the
         Bunge convention :cite:`rowenhorst2015consistent`.