+++ orix/plot/_util.py
@@ -41,12 +41,12 @@ class Arrow3D(FancyArrowPatch):
         return np.min(zs)
 
 
-def format_vector_labels(
+def format_labels(
     labels: np.ndarray,
     brackets: Tuple[str, str] = ("", ""),
     use_latex: bool = True,
 ) -> List[str]:
-    """Return formatted labels.
+    """Return formatted vector labels.
 
     This function is a convenient way to get nice labels when plotting
     vectors in the stereographic projection via
@@ -80,15 +80,15 @@ def format_vector_labels(
     >>> from orix.vector import Vector3d
     >>> v = Vector3d([[1, 1, 1], [-2, 0, 1], [4, 0, 0], [-4, 0, 0]])
     >>> v = v.reshape((2, 2))
-    >>> plot.format_vector_labels(v.data)
+    >>> plot.format_labels(v.data)
     ['$111$', '$\\bar{2}01$', '$400$', '$\\bar{4}00$']
-    >>> plot.format_vector_labels(v.data, ("[", "]"), use_latex=False)
+    >>> plot.format_labels(v.data, ("[", "]"), use_latex=False)
     ['[111]', '[-201]', '[400]', '[-400]']
-    >>> plot.format_vector_labels(v.data, ("{", "}"))
+    >>> plot.format_labels(v.data, ("{", "}"))
     ['$\\{111\\}$', '$\\{\\bar{2}01\\}$', '$\\{400\\}$', '$\\{\\bar{4}00\\}$']
-    >>> plot.format_vector_labels(v.data, ("{", "}"), use_latex=False)
+    >>> plot.format_labels(v.data, ("{", "}"), use_latex=False)
     ['{111}', '{-201}', '{400}', '{-400}']
-    >>> plot.format_vector_labels(v.data, ("<", ">"))
+    >>> plot.format_labels(v.data, ("<", ">"))
     ['$\\left<111\\right>$',
      '$\\left<\\bar{2}01\\right>$',
      '$\\left<400\\right>$',