+++ CHANGELOG.rst
@@ -11,8 +11,9 @@ Unreleased
 
 Added
 -----
+<<<<<<< HEAD
 - ``Vector3d.from_path_ends()`` class method to get vectors between two vectors.
-- Convenience function ``plot.format_vector_labels()`` to get a list of nicely formatted
+- Convenience function ``plot.format_labels()`` to get a list of nicely formatted vector
   labels to use when plotting vectors.
 - Two offsets in the stereographic coordinates (X, Y) can be given to
   ``StereographicPlot.text()`` to offset text coordinates.